requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
pydantic==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
jinja2==3.1.2

#redis==5.0.1  # Optional for caching
#twilio==8.14.0  # Optional for SMS
httpx==0.25.1  # For async HTTP requests
bcrypt==4.0.1